requirements.txt
google-ads>=24.0.0
anthropic>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
